test(deepseek_v4): cut the suite to a mergeable core
The V4 tests had grown to 51 across 1,831 lines, 37% of the repo's entire
GPU unit-test count for one model, where every other model gets between 2
and 8. The cause is structural: transformers.models.deepseek_v4 only exists
from 5.15 and the repo pins older, so the parity-against-HF archetype the
other models use is unavailable and the suite hand-wrote 45 substitutes.

Cut to 16 (13 GPU, 3 CPU), within a factor of two of nemotron_h, the most
complex precedent. What stays is the merge-critical core plus the tests born
red against bugs actually fixed on this branch: per-document compression,
sliding-window clipping at document boundaries, YaRN on the compressed
attention branch, the unprefixed on-disk key naming, and the legacy
compress_ratios translation. Dropped are closed-form properties of static
formulas that were never red, checks the whole-model forward subsumes,
degenerate-shape edge cases, and over-parametrization (compressor packing 9
cases to 2, vLLM RoPE 4 to 1). The strict xfail goes with them: it documented
that inject_prime_lm_head passes a 1-based arange the model rejects, which is
a bug report rather than a test.

test_packed_logits_match_unpacked becomes test_deepseek_v4, since with no
reference implementation available it is the only oracle covering the
assembled stack, and it now reads as the model's primary test the way
test_llama and test_nemotron_h do.

Fold deepseek_v4_helpers.py back into the two test modules. No other model
has a shared helpers module, and most of it died with the dropped tests.
_ATTN went with it: its keys were a strict subset of _MODEL's and
DeepseekV4Attention reads none of the fields that differed, so building the
per-mechanism layers from the one config is bit-for-bit a no-op, verified
over both layer types and four packings.

Confirmed the survivors still falsify: dropping the same_document term from
build_sliding_window_mask turns four of them red, test_deepseek_v4 included.
